# perf: don't deep copy the annotations in FromK8sObject