--- a/internal/util/objectmeta.go
+++ b/internal/util/objectmeta.go
@@ -1,8 +1,6 @@
 package util
 
 import (
-	"maps"
-
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -15,14 +13,14 @@ type K8sObjectInfo struct {
 	GroupVersionKind schema.GroupVersionKind
 }
 
+// FromK8sObject extracts information from a Kubernetes object.
+// It performs a shallow copy of object annotations so any modifications after
+// calling FromK8sObject will have an effect on the original object.
 func FromK8sObject(obj client.Object) K8sObjectInfo {
-	ret := K8sObjectInfo{
-		Name:        obj.GetName(),
-		Namespace:   obj.GetNamespace(),
-		Annotations: maps.Clone(obj.GetAnnotations()),
+	return K8sObjectInfo{
+		Name:             obj.GetName(),
+		Namespace:        obj.GetNamespace(),
+		Annotations:      obj.GetAnnotations(),
+		GroupVersionKind: obj.GetObjectKind().GroupVersionKind(),
 	}
-	if gvk := obj.GetObjectKind().GroupVersionKind(); gvk.String() != "" {
-		ret.GroupVersionKind = gvk
-	}
-	return ret
 }