**What this PR does / why we need it**:

No need to deep copy the annotations:

```
goos: darwin
goarch: arm64
pkg: github.com/kong/kubernetes-ingress-controller/v3/internal/util
                 │   old.txt    │              new.txt               │
                 │    sec/op    │   sec/op     vs base               │
FromK8sObject-10   202.30n ± 3%   44.08n ± 0%  -78.21% (p=0.008 n=5)

                 │  old.txt   │             new.txt              │
                 │    B/op    │   B/op    vs base                │
FromK8sObject-10   384.0 ± 0%   0.0 ± 0%  -100.00% (p=0.008 n=5)

                 │  old.txt   │              new.txt               │
                 │ allocs/op  │ allocs/op   vs base                │
FromK8sObject-10   3.000 ± 0%   0.000 ± 0%  -100.00% (p=0.008 n=5)
```