@randmonkey 2024-07-05
Can we guarantee that we never write the `annotations`? I think there is no need to write to it but we need to notice that.